ehkä korjattu boolean-vastausongelma

+++ start.py
@@ -116,8 +116,8 @@ def haeKysymys():
 # return oikeiden määrä?
 @app.route('/tarkistaVastaus', methods=['POST'])
 def tarkistaVastaus():
-    kysymys = request.json["kysymysID"]
-    vastaus = request.json['vastausID']
+    kysymys = (int)(request.json["kysymysID"])
+    vastaus = (int)(request.json['vastausID'])
 
     ov = haeKannasta(lambda c: hae_kysymys_ksm_ov(kysymys, c))[0][1]
     print(ov)
@@ -125,10 +125,7 @@ def tarkistaVastaus():
         onko_oikein = vastaus == ov
 
     else: onko_oikein = haeKannasta(lambda c: tarkista_onko_oikein(vastaus, c))[0][0]
-    
-    # hae tässä datasta haluttu tieto, eli oikeiden vastausten taulukko
-    # tarkista
-    # palauta mitä?
+
     return (str)(onko_oikein), 200
